+++ src/containers/array.c
@@ -217,7 +217,7 @@ void array_container_andnot(const array_container_t *array_1,
     if (out->capacity < array_1->cardinality)
         array_container_grow(out, array_1->cardinality, false);
 #ifdef CROARING_IS_X64
-    if(( croaring_avx2() ) && (out != array_1) && (out != array_2)) {
+    if(( croaring_hardware_support() & ROARING_SUPPORTS_AVX2 ) && (out != array_1) && (out != array_2)) {
       out->cardinality =
           difference_vector16(array_1->array, array_1->cardinality,
                             array_2->array, array_2->cardinality, out->array);
@@ -248,7 +248,7 @@ void array_container_xor(const array_container_t *array_1,
     }
 
 #ifdef CROARING_IS_X64
-    if( croaring_avx2() ) {
+    if( croaring_hardware_support() & ROARING_SUPPORTS_AVX2 ) {
       out->cardinality =
         xor_vector16(array_1->array, array_1->cardinality, array_2->array,
                      array_2->cardinality, out->array);
@@ -297,7 +297,7 @@ void array_container_intersection(const array_container_t *array1,
             array2->array, card_2, array1->array, card_1, out->array);
     } else {
 #ifdef CROARING_IS_X64
-       if( croaring_avx2() ) {
+       if( croaring_hardware_support() & ROARING_SUPPORTS_AVX2 ) {
         out->cardinality = intersect_vector16(
             array1->array, card_1, array2->array, card_2, out->array);
        } else {
@@ -325,7 +325,7 @@ int array_container_intersection_cardinality(const array_container_t *array1,
                                                    array1->array, card_1);
     } else {
 #ifdef CROARING_IS_X64
-    if( croaring_avx2() ) {
+    if( croaring_hardware_support() & ROARING_SUPPORTS_AVX2 ) {
         return intersect_vector16_cardinality(array1->array, card_1,
                                               array2->array, card_2);
     } else {
@@ -371,7 +371,7 @@ void array_container_intersection_inplace(array_container_t *src_1,
             src_2->array, card_2, src_1->array, card_1, src_1->array);
     } else {
 #ifdef CROARING_IS_X64
-        if (croaring_avx2()) {
+        if (croaring_hardware_support() & ROARING_SUPPORTS_AVX2) {
             src_1->cardinality = intersect_vector16_inplace(
                 src_1->array, card_1, src_2->array, card_2);
         } else {
